Completed Orders

+++ Project/Loothub/src/main/java/com/company/springboot/repository/IProductRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IProductRepository extends JpaRepository<Product, Long>{
+public interface IProductRepository extends JpaRepository<Product, Integer>{
     
 }
